Added test to entiresuite
